split up tests for mapchete.io.vector into submodules
